requirements.txt
Pillow==6.1.0
pkg-resources==0.0.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
vext==0.7.3
vext.gi==0.7.0
